Use CommonJS export instead of Babel for better non-Babel compatibility
Default exports in Babel export on the default property. Combined with
UMD this would lead to prescribe being available in the browser as
‘default’. CommonJS users not using Babel would have to import prescribe
as `var prescribe = require(‘prescribe’).default`.

By using a standard CommonJS export combined with the webpack
`output.library` option everything works as expected everywhere,
and prescribe will be available in browser contexts as `Prescribe`.