Add test:debug and specify +1 browsers
Sauce allows an empty string to mean latest version of this browser. I've specified the latest and the current previous version here in an impl of the `+1` method for testing (along w/ IE8+). I set windows for most as that seemed to be the most stable OS on Sauce.

I tried the `latest-1` method specified on sauce's site, but it consistently failed.