fix(HtmlParser): return null for unclosed script and style tags
In the original version of HtmlParser embedded in postscribe, we would return null for script and
style tokens.  This behaviour appears to have changed in the extraction and we should revert to that
earlier behaviour since it makes no sense to return just the start tag for script and style.

Fixes #35